refactor(ai): move retry utilities into dedicated module
